+++ analyser/application/repository_analysis/tests/test_languages.py
@@ -1,82 +1,14 @@
 from unittest.mock import MagicMock, patch
 
 from application.repository_analysis.languages import (
-    add_languages_sloc_table,
-    catalogue_repository,
-    check_for_excluded_dirs,
-    clone_repo,
     count_files_per_language,
     determine_file_language,
+    add_to_catalogued_files,
+    select_best_match,
+    language_match,
 )
 
 FILE_PATH = "application.repository_analysis.languages"
-
-
-@patch(f"{FILE_PATH}.count_files_per_language")
-@patch(f"{FILE_PATH}.catalogue_repository")
-@patch(f"{FILE_PATH}.get_languages")
-@patch(f"{FILE_PATH}.clone_repo")
-def test_add_languages_sloc_table(
-    mock_clone_repo: MagicMock,
-    mock_get_languages: MagicMock,
-    mock_catalogue_repository: MagicMock,
-    mock_count_files_per_language: MagicMock,
-) -> None:
-    # Arrange
-    markdown_file = MagicMock()
-    mock_count_files_per_language.return_value = {"Python": 100, "Javascript": 200}
-    # Act
-    add_languages_sloc_table(markdown_file)
-    # Assert
-    markdown_file.new_table.assert_called_once_with(
-        columns=2, rows=3, text=["Language", "File Count", "Python", 100, "Javascript", 200]
-    )
-    mock_clone_repo.assert_called_once_with("JackPlowman", "github-stats")
-    mock_get_languages.assert_called_once_with()
-    mock_catalogue_repository.assert_called_once_with(mock_clone_repo.return_value, mock_get_languages.return_value)
-    mock_count_files_per_language.assert_called_once_with(mock_catalogue_repository.return_value)
-
-
-@patch(f"{FILE_PATH}.Path")
-@patch(f"{FILE_PATH}.Repo")
-def test_clone_repo(mock_repo: MagicMock, mock_path: MagicMock) -> None:
-    # Arrange
-    mock_path.exists.return_value = False
-    # Act
-    clone_repo("JackPlowman", "github-stats")
-    # Assert
-    mock_repo.clone_from.assert_called_once_with(
-        "https://github.com/JackPlowman/github-stats.git", mock_path.return_value
-    )
-
-
-@patch(f"{FILE_PATH}.Path")
-@patch(f"{FILE_PATH}.Repo")
-def test_clone_repo_exists(mock_repo: MagicMock, mock_path: MagicMock) -> None:
-    # Arrange
-    mock_path.exists.return_value = True
-    # Act
-    clone_repo("JackPlowman", "github-stats")
-    # Assert
-    mock_repo.clone_from.assert_not_called()
-
-
-@patch(f"{FILE_PATH}.determine_file_language")
-@patch(f"{FILE_PATH}.check_for_excluded_dirs")
-@patch(f"{FILE_PATH}.Path")
-def test_catalogue_repository(
-    mock_path: MagicMock, mock_check_for_excluded_dirs: MagicMock, mock_determine_file_language: MagicMock
-) -> None:
-    # Arrange
-    mock_languages = []
-    file_path = "application/repos/github-stats"
-    mock_path.return_value.walk.return_value = [(mock_path, [], ["file1.py", "file2.py"])]
-    mock_check_for_excluded_dirs.return_value = False
-    mock_determine_file_language.return_value = expected = {"Python": ["file1.py", "file2.py"]}
-    # Act
-    response = catalogue_repository(file_path, mock_languages)
-    # Assert
-    assert response == expected
 
 
 def test_determine_file_language() -> None:
@@ -93,14 +25,17 @@ def test_determine_file_language() -> None:
     assert response == {mock_language.name: [f"{mock_path}/{mock_file_name}"]}
 
 
-def test_check_for_excluded_dirs() -> None:
+def test_add_to_catalogued_files() -> None:
     # Arrange
-    mock_path = MagicMock()
+    mock_catalogued_files = "test.py"
+    mock_language = MagicMock()
+    mock_language.extensions = [".py"]
+    mock_language.name = "Python"
+    catalogued_files = {"Markdown": ["file1.md", "file2.md"]}
     # Act
-    check_for_excluded_dirs(mock_path, ["excluded"])
+    response =add_to_catalogued_files(mock_catalogued_files, catalogued_files, mock_language)
     # Assert
-
-    mock_path.assert_not_called()
+    assert response == catalogued_files | {mock_language.name: ["test.py"]}
 
 
 def test_count_files_per_language() -> None:
@@ -110,3 +45,34 @@ def test_count_files_per_language() -> None:
     response = count_files_per_language(file_types)
     # Assert
     assert response == {"Python": 2}
+
+def test_select_best_match() -> None:
+    # Arrange
+    mock_file_name = "file.py"
+    mock_possible_matches = [MagicMock(), MagicMock()]
+    mock_possible_matches[0].name = "Python"
+    mock_possible_matches[1].name = "Javascript"
+    # Act
+    response = select_best_match(mock_file_name, mock_possible_matches)
+    # Assert
+    assert response == mock_possible_matches[0]
+
+def test_language_match() -> None:
+    # Arrange
+    mock_language = MagicMock()
+    mock_language.name = "Python"
+    mock_prioritised_languages = ["Python", "Javascript"]
+    # Act
+    response = language_match(mock_language, mock_prioritised_languages)
+    # Assert
+    assert response is True
+
+def test_language_match__no_match() -> None:
+    # Arrange
+    mock_language = MagicMock()
+    mock_language.name = "Java"
+    mock_prioritised_languages = ["Python", "Javascript"]
+    # Act
+    response = language_match(mock_language, mock_prioritised_languages)
+    # Assert
+    assert response is False
